new version

+++ src/watermelon/sim/Pair.java
@@ -3,12 +3,12 @@ package watermelon.sim;
 import watermelon.sim.Pair;
 
 public class Pair {
-    public int x;
-    public int y;
+    public double x;
+    public double y;
 
-    public Pair() { x = 0; y = 0; }
+    public Pair() { x = 0.0; y = 0.0; }
 
-    public Pair(int xx, int yy) {
+    public Pair(double xx, double yy) {
         x = xx;
         y = yy;
     }
