+++ src/watermelon/sim/Point.java
@@ -2,11 +2,11 @@ package watermelon.sim;
 import watermelon.sim.Point;
 
 public class Point {
-    public int x;
-    public int y;
+    public double x;
+    public double y;
     public boolean tree;
     //public int owner;
-    public Point(int xx, int yy, boolean t) {
+    public Point(double xx, double yy, boolean t) {
         x = xx;
         y = yy;
         tree =t;
