+++ src/watermelon/sim/seed.java
@@ -4,13 +4,13 @@ import java.util.ArrayList;
 
 
 public class seed {
-    public int x;
-    public int y;
+    public double x;
+    public double y;
     public boolean tetraploid;
     
-    public seed() { x = 0; y = 0; tetraploid = false; }
+    public seed() { x = 0.0; y = 0.0; tetraploid = false; }
 
-    public seed(int xx, int yy, boolean tetra) {
+    public seed(double xx, double yy, boolean tetra) {
         x = xx;
         y = yy;
         tetraploid = tetra;
